Fix integer issues in color math.

+++ bibliopixel/colors/arithmetic.py
@@ -7,4 +7,4 @@ def color_blend(a, b):
 
 def color_scale(color, level):
     """Scale RGB tuple by level, 0 - 256"""
-    return tuple([(i * level) >> 8 for i in list(color)])
+    return tuple([int(i * level) >> 8 for i in list(color)])
